Update src/js/me-mediaelements.js
Fixed setSrc for YouTube videos which needed to use a YouTube API specific call. Changed it to cueVideoById, with a call to a newly defined mejs.Utility function called getYouTubeVideoId, which extracts the YouTube video id from the URL. The player is also stopped before doing so.

When setVideoSize was called, this.pluginElement was often null for YouTube, so a check for this was placed around the check for this.pluginElement.style.